Module backup::endpoint

Structs

BackupRange

ConfigManager

ControlThreadPool

Endpoint
The endpoint of backup.

LimitedStorage

Progress
The progress of a backup task

Request

Task
Backup Task.

Constants

BACKUP_BATCH_LIMIT

IDLE_THREADPOOL_DURATION

WORKER_TAKE_RANGE

Functions

backup_file_name
Construct an backup file name based on the given store id and region. A name consists with three parts: store id, region_id and a epoch version.

get_max_start_key
Get the max start key from the given start_key and Region's start key.

get_min_end_key
Get the min end key from the given end_key and Region's end key.

to_sst_compression_type