Function backup::endpoint::backup_file_name

fn backup_file_name(
    store_id: u64, 
    region: &Region, 
    key: Option<String>
) -> String

Construct an backup file name based on the given store id, region, range start key and local unix timestamp. A name consists with five parts: store id, region_id, a epoch version, the hash of range start key and timestamp. range start key is used to keep the unique file name for file, to handle different tables exists on the same region. local unix timestamp is used to keep the unique file name for file, to handle receive the same request after connection reset.